goods PostMan测试

+++ goods/src/main/java/cn/edu/xmu/goods/model/bo/FloatPrice.java
@@ -64,7 +64,7 @@ public class FloatPrice implements VoObject {
 
     private LocalDateTime endTime;
 
-    private Long quantity;
+    private Integer quantity;
 
     private Long createdBy;
 
@@ -82,7 +82,7 @@ public class FloatPrice implements VoObject {
         this.setActivityPrice(po.getActivityPrice());
         this.setBeginTime(po.getBeginTime());
         this.setEndTime(po.getEndTime());
-        this.setQuantity(po.getQuantity().longValue());
+        this.setQuantity(po.getQuantity());
         this.setCreatedBy(po.getCreatedBy());
         this.setInvalidBy(po.getInvalidBy());
         this.setValid(po.getValid());
@@ -97,7 +97,7 @@ public class FloatPrice implements VoObject {
         po.setActivityPrice(this.getActivityPrice());
         po.setBeginTime(this.getBeginTime());
         po.setEndTime(this.getEndTime());
-        po.setQuantity(this.getQuantity().intValue());
+        po.setQuantity(this.getQuantity());
         po.setCreatedBy(this.getCreatedBy());
         po.setInvalidBy(this.getInvalidBy());
         po.setValid(this.getValid());
